% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{get_lambda}
\alias{get_lambda}
\title{Return the secret lambda from a private key}
\usage{
get_lambda(private_key, ...)
}
\arguments{
\item{private_key}{a \link{PaillierPrivateKey}.}

\item{...}{unused.}
}
\value{
a \link[gmp:bigz]{gmp::bigz} value.
}
\description{
Return the secret lambda from a private key
}
