% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{get_private_key}
\alias{get_private_key}
\title{Return the private key from a key pair}
\usage{
get_private_key(keypair, ...)
}
\arguments{
\item{keypair}{a \link{PaillierKeyPair}.}

\item{...}{unused.}
}
\value{
a \link{PaillierPrivateKey}.
}
\description{
Return the private key from a key pair
}
