% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{PaillierCiphertext}
\alias{PaillierCiphertext}
\title{A Paillier ciphertext}
\usage{
PaillierCiphertext(value = NULL, pubkey = PaillierPublicKey())
}
\arguments{
\item{value}{the encrypted big-integer value.}

\item{pubkey}{the \link{PaillierPublicKey} under which it was encrypted.}
}
\description{
Wraps the encrypted big-integer value together with the public key
it was encrypted under. Two ciphertexts encrypted under the same
public key can be combined with \code{+} and \code{-}; a ciphertext can be
multiplied by a cleartext integer with \code{*}.
}
