% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{PaillierPrivateKey}
\alias{PaillierPrivateKey}
\title{Paillier private key}
\usage{
PaillierPrivateKey(lambda, pubkey)
}
\arguments{
\item{lambda}{the secret lambda.}

\item{pubkey}{the matching \link{PaillierPublicKey}.}
}
\description{
Holds the secret \code{lambda} and a cached value \code{x} used during
decryption, together with a reference to the matching public key.
}
