% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homomorpheR.R
\name{random.bigz}
\alias{random.bigz}
\title{Random big integer}
\usage{
random.bigz(nBits)
}
\arguments{
\item{nBits}{number of bits, which must be a multiple of 8 (not
checked, for efficiency).}
}
\value{
a \link[gmp:bigz]{gmp::bigz} value.
}
\description{
Returns a random big integer using the cryptographically secure
generator from the \code{sodium} package.
}
