% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{encrypt}
\alias{decrypt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{openfhe.R}{\code{\link[openfhe.R:decrypt]{decrypt()}}, \code{\link[openfhe.R:encrypt]{encrypt()}}}
}}

