% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_legacy.R
\name{round_robin_chain}
\alias{round_robin_chain}
\title{Wire a master and a list of sites into a round-robin chain}
\usage{
round_robin_chain(master, sites)
}
\arguments{
\item{master}{a \link{Master}.}

\item{sites}{a list of \link{Site}s.}
}
\value{
the master, invisibly.
}
\description{
Sets \code{master -> sites[[1]] -> sites[[2]] -> ... -> sites[[n]] -> master}
and broadcasts the master's public key to every site. After this
call, \code{\link[=run_round_robin]{run_round_robin()}} can drive an iteration of the protocol.
}
\details{
Part of the frozen Paillier-era legacy surface; the supported
topology is \code{\link[=set_workers]{set_workers()}} + \code{\link[=master_aggregate]{master_aggregate()}}.
}
