% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_legacy.R
\name{add_site}
\alias{add_site}
\title{Add a site to a non-cooperating party}
\usage{
add_site(ncp, ...)
}
\arguments{
\item{ncp}{an \link{NCParty}.}

\item{...}{method-specific arguments. The NCParty method takes a
single \link{Site}.}
}
\description{
Part of the frozen Paillier-era legacy surface.
}
