% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt a value under a Paillier public key}
\usage{
encrypt(public_key, ...)
}
\arguments{
\item{public_key}{a \link{PaillierPublicKey}.}

\item{...}{method-specific arguments. The Paillier method takes a
single cleartext value \code{x} (integer or \link[gmp:bigz]{gmp::bigz}).}
}
\value{
a \link{PaillierCiphertext}.
}
\description{
Encrypt a value under a Paillier public key
}
