% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homomorpheR.R
\docType{package}
\name{homomorpheR}
\alias{homomorpheR-package}
\alias{homomorpheR}
\title{homomorpheR: Homomorphic computations in R}
\description{
\code{homomorpheR} provides homomorphic encryption schemes for
privacy-preserving distributed computations: applications of the
sort implemented in package \code{distcomp}. The Paillier cryptosystem
is implemented natively in R via the \code{gmp} package; CKKS, BFV, BGV,
and FHEW/TFHE schemes are available through the \code{openfhe} package.
}
\details{
Encrypted values are wrapped in \link{PaillierCiphertext} objects so that
R's arithmetic operators dispatch to the homomorphism. Use
\code{\link[=paillier_keypair]{paillier_keypair()}} to generate keys, \code{\link[=encrypt]{encrypt()}} to encrypt, and
\code{\link[=decrypt]{decrypt()}} to recover the result.

For a quick overview, see the package vignettes.
}
\examples{
\dontrun{
keys <- paillier_keypair(1024)
encrypt_decrypt <- function(x) decrypt(get_private_key(keys),
                                       encrypt(keys@pubkey, x))
a <- gmp::as.bigz(1273849)
identical(a + 10L, encrypt_decrypt(a + 10L))
}
}
\references{
\href{https://en.wikipedia.org/wiki/Homomorphic_encryption}{Homomorphic Encryption}

\href{https://mhe.github.io/jspaillier/}{Paillier Encryption}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bnaras.github.io/homomorpheR/}
  \item Report bugs at \url{https://github.com/bnaras/homomorpheR/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stat.Stanford.EDU}

}
