% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{make_ckks_master}
\alias{make_ckks_master}
\title{Construct a CKKS-backed master}
\usage{
make_ckks_master(name, crypto_context, keypair)
}
\arguments{
\item{name}{short identifier shown in printed output.}

\item{crypto_context}{an \code{openfhe.R} \code{CryptoContext} configured for
CKKS.}

\item{keypair}{an \code{openfhe.R} \code{KeyPair}.}
}
\value{
a \link{CKKSMaster}.
}
\description{
Construct a CKKS-backed master
}
