% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{make_master}
\alias{make_master}
\title{Construct a Paillier-backed master}
\usage{
make_master(name, keypair, den = gmp::as.bigq(2)^256)
}
\arguments{
\item{name}{short identifier shown in printed output.}

\item{keypair}{a \link{PaillierKeyPair}.}

\item{den}{a \link[gmp:bigq]{gmp::bigq} denominator used to scale fractional parts
when encrypting real numbers via \code{\link[=encrypt_real]{encrypt_real()}}.}
}
\value{
a \link{PaillierMaster}.
}
\description{
Construct a Paillier-backed master
}
