% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{make_ncparty}
\alias{make_ncparty}
\title{Construct an \link{NCParty}}
\usage{
make_ncparty(name, number)
}
\arguments{
\item{name}{short identifier shown in printed output.}

\item{number}{which share this NCP receives, \code{1} or \code{2}.}
}
\value{
an \link{NCParty}.
}
\description{
Construct an \link{NCParty}
}
