% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\name{PaillierKeyPair}
\alias{PaillierKeyPair}
\title{Paillier key pair}
\usage{
PaillierKeyPair(pubkey = PaillierPublicKey(), privkey = PaillierPrivateKey())
}
\arguments{
\item{pubkey}{a \link{PaillierPublicKey}.}

\item{privkey}{a \link{PaillierPrivateKey}.}
}
\description{
A matched pair of public and private keys. Use \code{\link[=paillier_keypair]{paillier_keypair()}}
to generate one.
}
