% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\docType{class}
\name{PaillierKeyPair}
\alias{PaillierKeyPair}
\title{Construct a Paillier public and private key pair given a fixed number of bits}
\format{An \code{\link{R6Class}} generator object}
\usage{
PaillierKeyPair
}
\description{
Construct a Paillier public and private key pair given a fixed number of bits
}
\section{Fields}{

\describe{
\item{\code{pubkey}}{the Paillier public key}
}}
\section{Methods}{

\describe{
  \item{\code{PaillierKeyPair$new(modulusBits)}}{Create a new private key with specified
        number of modulus bits}
  \item{\code{PaillierKeyPair$getPrivateKey()}}{Return the private key}
}
}
\examples{
keys <- PaillierKeyPair$new(1024)
keys$pubkey
keys$getPrivateKey()

}
\seealso{
\code{PaillierPublicKey} and \code{PaillierPrivateKey}
}
\keyword{datasets}

