% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paillier.R
\docType{class}
\name{PaillierPrivateKey}
\alias{PaillierPrivateKey}
\title{Construct a Paillier private key with the given secret and a public key}
\format{An \code{\link{R6Class}} generator object}
\usage{
PaillierPrivateKey
}
\description{
Construct a Paillier private key with the given secret and a public key
}
\section{Fields}{

\describe{
\item{\code{pubkey}}{the Paillier public key}
}}
\section{Methods}{

\describe{
  \item{\code{PaillierPrivateKey$new(lambda, pubkey)}}{Create a new private key with given secret
        \code{lambda} and the public key}
  \item{\code{PaillierPrivateKey$getLambda()}}{Return the secret \code{lambda}}
  \item{\code{PaillierPrivateKey$decrypt(c)}}{Decrypt a message. The value \code{c} should be an
        encrypted value}
}
}
\seealso{
\code{PaillierPublicKey} which goes hand-in-hand with this object
}
\keyword{datasets}

