% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homomorpheR.R
\name{random.bigz}
\alias{random.bigz}
\title{Return a random big number using the cryptographically secure random number generator
from in the \code{sodium} package.}
\usage{
random.bigz(nBits)
}
\arguments{
\item{nBits, }{the number of bits, which must be a multiple of 8, is not checked for efficiency.}
}
\description{
Return a random big number using the cryptographically secure random number generator
from in the \code{sodium} package.
}

