% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{set_next_site}
\alias{set_next_site}
\title{Wire one site's \code{next_site} to another}
\usage{
set_next_site(obj, ...)
}
\arguments{
\item{obj}{a \link{Site} or \link{Master}.}

\item{...}{method-specific arguments. The Site/Master methods take
a single \code{next_site}.}
}
\description{
Wire one site's \code{next_site} to another
}
