% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{set_public_key}
\alias{set_public_key}
\title{Distribute the public key from the master to a downstream actor}
\usage{
set_public_key(obj, ...)
}
\arguments{
\item{obj}{an \link{NCParty} or \link{Site} to receive the key.}

\item{...}{method-specific arguments. The Site/NCParty methods
take a single \link{PaillierPublicKey} \code{pubkey}.}
}
\description{
Distribute the public key from the master to a downstream actor
}
