% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{set_public_key}
\alias{set_public_key}
\title{Distribute the public key from the master to a downstream actor}
\usage{
set_public_key(obj, ...)
}
\arguments{
\item{obj}{a \link{Site} (or legacy \link{NCParty}) to receive the key.}

\item{...}{method-specific arguments. The methods take a single
public key \code{pubkey} of the master's backend type.}
}
\description{
Distribute the public key from the master to a downstream actor
}
