% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{set_workers}
\alias{set_workers}
\title{Wire a master to a flat list of workers}
\usage{
set_workers(master, workers)
}
\arguments{
\item{master}{a \link{Master}.}

\item{workers}{a list of worker \link{Site}s.}
}
\value{
the master, invisibly.
}
\description{
Stashes the workers in the master's state and broadcasts the
master's public key to each worker. After this call,
\code{\link[=run_master_worker]{run_master_worker()}} can drive an iteration of the protocol.
}
\details{
Use this for the realistic master/worker (star) topology that
distcomp- and DataSHIELD-style federated analyses follow. For the
legacy Paillier round-robin idiom, use \code{\link[=round_robin_chain]{round_robin_chain()}} instead.
}
